Rename reactive.py to react.py and get option
